+++ src/forms/field/field.controller.js
@@ -7,8 +7,9 @@ class FieldController {
     if (changes.initialField) {
       this.field = copyJSON(this.initialField);
 
-      this.control = this.field.control ? this.field.control
-        : this.RequirementsService.getControlType(changes.initialField.currentValue);
+      this.control = this.RequirementsService.getControlType(
+        changes.initialField.currentValue
+      );
 
       // TODO we should probably do this at fieldset level, so the model is available
       if (this.field.valuesAsync) {