+++ src/forms/fieldset/fieldset.demo.js
@@ -227,6 +227,16 @@ function fieldsetDocsController() {
           }
         ]
       },
+      checkboxGroup: {
+        title: 'Checkbox Group',
+        type: 'string',
+        control: 'checkbox-group',
+        width: 'md',
+        items: {
+          enum: [1, 2],
+          values: [{ value: 1, label: 'One' }, { value: 2, label: 'Two' }]
+        }
+      },
       password: {
         title: 'Password',
         type: 'string',
@@ -298,6 +308,7 @@ function fieldsetDocsController() {
     checkbox: true,
     radio: '2',
     password: 'qwerty',
-    telephone: '+441234567890'
+    telephone: '+441234567890',
+    checkboxGroup: '["1", "2"]'
   };
 }